Update acanfd library to v2.1.16
