+++ Software/src/lib/pierremolinaro-ACAN2517FD/ACAN2517FDFilters.h
@@ -1,27 +1,27 @@
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 // An utility class for:
 //   - ACAN2517FD CAN driver for MCP2517FD (CANFD mode)
 // by Pierre Molinaro
 // https://github.com/pierremolinaro/acan2517FD
 //
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 #ifndef ACAN2517FD_FILTERS_CLASS_DEFINED
 #define ACAN2517FD_FILTERS_CLASS_DEFINED
 
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
-#include "CANFDMessage.h"
+#include <ACAN2517FD_CANFDMessage.h>
 
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 //  ACAN2517FDFilters class
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 class ACAN2517FDFilters {
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   EMBEDDED CLASS
-//······················································································································
+//------------------------------------------------------------------------------
 
   private: class Filter {
     public: Filter * mNextFilter ;
@@ -43,9 +43,9 @@ class ACAN2517FDFilters {
     private: Filter & operator = (const Filter &) ;
   } ;
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   ENUMERATED TYPE
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: typedef enum {
       kFiltersOk,
@@ -58,15 +58,15 @@ class ACAN2517FDFilters {
       kInconsistencyBetweenMaskAndAcceptance
   } FilterStatus ;
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   CONSTRUCTOR
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: ACAN2517FDFilters (void) {}
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   DESTRUCTOR
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: ~ ACAN2517FDFilters (void) {
     while (mFirstFilter != NULL) {
@@ -76,9 +76,9 @@ class ACAN2517FDFilters {
     }
   }
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   RECEIVE FILTERS
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: void appendPassAllFilter (const ACANFDCallBackRoutine inCallBackRoutine) {  // Accept any frame
     Filter * f = new Filter (0, 0, inCallBackRoutine) ;
@@ -91,7 +91,7 @@ class ACAN2517FDFilters {
     mFilterCount += 1 ;
   }
 
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: void appendFormatFilter (const tFrameFormat inFormat, // Accept any identifier
                                    const ACANFDCallBackRoutine inCallBackRoutine) {
@@ -107,7 +107,7 @@ class ACAN2517FDFilters {
     mFilterCount += 1 ;
   }
 
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: void appendFrameFilter (const tFrameFormat inFormat,
                                   const uint32_t inIdentifier,
@@ -141,7 +141,7 @@ class ACAN2517FDFilters {
     mFilterCount += 1 ;
   }
 
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: void appendFilter (const tFrameFormat inFormat,
                              const uint32_t inMask,
@@ -196,9 +196,9 @@ class ACAN2517FDFilters {
     mFilterCount += 1 ;
   }
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   ACCESSORS
-//······················································································································
+//------------------------------------------------------------------------------
 
   public: FilterStatus filterStatus (void) const { return mFilterStatus ; }
 
@@ -206,9 +206,9 @@ class ACAN2517FDFilters {
 
   public: uint8_t filterCount (void) const { return mFilterCount ; }
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   PRIVATE PROPERTIES
-//······················································································································
+//------------------------------------------------------------------------------
 
   private: uint8_t mFilterCount = 0 ;
   private: Filter * mFirstFilter = NULL ;
@@ -216,23 +216,23 @@ class ACAN2517FDFilters {
   private: FilterStatus mFilterStatus = kFiltersOk ;
   private: uint8_t mFilterErrorIndex = 0 ;
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   NO COPY
-//······················································································································
+//------------------------------------------------------------------------------
 
   private: ACAN2517FDFilters (const ACAN2517FDFilters &) = delete ;
   private: ACAN2517FDFilters & operator = (const ACAN2517FDFilters &) = delete ;
 
-//······················································································································
+//------------------------------------------------------------------------------
 //   Friend
-//······················································································································
+//------------------------------------------------------------------------------
 
   friend class ACAN2517FD ;
 
-//······················································································································
+//------------------------------------------------------------------------------
 
 } ;
 
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 #endif
