+++ Software/src/lib/pierremolinaro-ACAN2517FD/ACAN2517FD_CANFDMessage.h
@@ -1,21 +1,21 @@
-//-----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 // Generic CANFD Message
 // by Pierre Molinaro
 //
 // https://github.com/pierremolinaro/acan2517FD
 //
-//-----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 #ifndef GENERIC_CANFD_MESSAGE_DEFINED
 #define GENERIC_CANFD_MESSAGE_DEFINED
 
-//-----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
-#include "CANMessage.h"
+#include <ACAN2517FD_CANMessage.h>
 
-//-----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 //    CANFDMessage class
-//-----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 // Note that "len" field contains the actual length, not its encoding in CANFD frames
 // Valid values are: 0, 1, ..., 8, 12, 16, 20, 24, 32, 48, 64.
 // Having other values is an error that prevents frame to be sent by tryToSend
@@ -23,9 +23,9 @@
 
 class CANFDMessage {
 
-//·············································································
-//   Constructors
-//·············································································
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+  //   Constructors
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
   public : CANFDMessage (void) :
   id (0),  // Frame identifier
@@ -36,7 +36,7 @@ class CANFDMessage {
   data () {
   }
 
-//·············································································
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
   public : CANFDMessage (const CANMessage & inMessage) :
   id (inMessage.id),  // Frame identifier
@@ -48,9 +48,9 @@ class CANFDMessage {
     data64 [0] = inMessage.data64 ;
   }
 
-//·············································································
-//   Enumerated Type
-//·············································································
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+  //   Enumerated Type
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
   public: typedef enum : uint8_t {
     CAN_REMOTE,
@@ -59,9 +59,9 @@ class CANFDMessage {
     CANFD_WITH_BIT_RATE_SWITCH
   } Type ;
 
-//·············································································
-//   Properties
-//·············································································
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+  //   Properties
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
   public : uint32_t id ;  // Frame identifier
   public : bool ext ; // false -> base frame, true -> extended frame
@@ -80,9 +80,9 @@ class CANFDMessage {
     uint8_t  data      [64] ;
   } ;
 
-//·············································································
-//   Methods
-//·············································································
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+  //   Methods
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
   public: void pad (void) {
     uint8_t paddedLength = len ;
@@ -107,7 +107,7 @@ class CANFDMessage {
     }
   }
 
-//·············································································
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
   public: bool isValid (void) const {
     if ((type == CAN_REMOTE) || (type == CAN_DATA)) { // Remote frame
@@ -121,14 +121,14 @@ class CANFDMessage {
     }
   }
 
-//·············································································
+  // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 } ;
 
-//-----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 typedef void (*ACANFDCallBackRoutine) (const CANFDMessage & inMessage) ;
 
-//-----------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 #endif
