+++ Software/src/lib/pierremolinaro-ACAN2517FD/ACAN2517FD_CANMessage.h
@@ -1,4 +1,4 @@
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 // Generic CAN Message
 // by Pierre Molinaro
 //
@@ -8,16 +8,16 @@
 // https://github.com/pierremolinaro/acan2517
 // https://github.com/pierremolinaro/acan2517FD
 //
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 #ifndef GENERIC_CAN_MESSAGE_DEFINED
 #define GENERIC_CAN_MESSAGE_DEFINED
 
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 #include <Arduino.h>
 
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 class CANMessage {
   public : uint32_t id = 0 ;  // Frame identifier
@@ -38,12 +38,12 @@ class CANMessage {
   } ;
 } ;
 
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 typedef enum {kStandard, kExtended} tFrameFormat ;
 typedef enum {kData, kRemote} tFrameKind ;
 typedef void (*ACANCallBackRoutine) (const CANMessage & inMessage) ;
 
-//----------------------------------------------------------------------------------------------------------------------
+//------------------------------------------------------------------------------
 
 #endif
