+++ Software/src/lib/pierremolinaro-ACAN2517FD/library.properties
@@ -1,0 +1,9 @@
+name=ACAN2517FD
+version=2.1.16
+author=Pierre Molinaro
+maintainer=Pierre Molinaro <Pierre.Molinaro@pcmolinaro.name>
+sentence=Driver for MCP2517FD and MCP2518FD CAN Controller (CAN FD mode)
+paragraph=This library is an Arduino CAN network driver for the MCP2517FD, the MCP2518FD and the MCP251863 CAN Controller, in CAN FD mode. Compatible with ACAN, ACAN2515, ACAN2517 libraries, with ACAN_T4 library from version 2.1.0. Default configuration sends and receives any frame – no default filter to provide. Reception filters (up to 32) can be easily defined. Compatible with ESP32 from version 1.1.0.
+category=Communication
+url=https://github.com/pierremolinaro/acan2517FD
+architectures=*
